Added full map view & changes
Added a full map view with its own camera, render texture and RawImage on the UI canvas. Also changed player speed and camera, changed the player prefab to include the pausepanel and comented on more code. Need to fix pause panel not interacting and change the big map raw image to the correct size.